refactor: send suggestions to supabase
